Misc changes (#39)
* SLUS-00868.cht

added enemies swap codes for go between default cars to hotrod non selectable versions for vigilante 8 second offense

* vigilante 8 and v8 2nd offense widescreen fix

added the widescreen fix since the game cut short the levels at the left and right. this expands the limits a bit.

* v8 and v8 2nd offense fix

fixing the missing code for the vigilante 8 and second offense that i missed when exported last time to my widescreen fix

* digimon tamers battle evolution codes fix

added the first line data the codes need to work properly

* Update SCUS-94560.cht

* Delete cheats/SLUS-00868(2).cht

deleted the file since somehow its already on the cht database of the release